requirements.txt
Django==4.2.2
django-crispy-forms
django-extensions
django-phonenumber-field
django-scheduler
num2words
Pillow
sentry-sdk
python-dotenv
crispy_bootstrap3
django_cleanup
phonenumbers
PyJWT
twilio
pygsheets
qrcode
sorl-thumbnail
gunicorn
django-googledrive-storage==1.6.0
google-api-python-client==2.90.0
